Extract waitForUrl function
