+++ src/index.js
@@ -1,23 +1,8 @@
 import waitForDeployCreation from './wait-for-deploy-creation'
 import waitForReadiness from './wait-for-readiness'
+import waitForUrl from './wait-for-url'
 import * as core from '@actions/core'
 import * as github from '@actions/github'
-import axios from 'axios'
-
-const waitForUrl = async (url, MAX_TIMEOUT) => {
-  const iterations = MAX_TIMEOUT / 3
-  for (let i = 0; i < iterations; i++) {
-    try {
-      await axios.get(url)
-
-      return
-    } catch (e) {
-      console.log(`URL ${url} unavailable, retrying...`)
-      await new Promise((r) => setTimeout(r, 3000))
-    }
-  }
-  core.setFailed(`Timeout reached: Unable to connect to ${url}`)
-}
 
 const run = async () => {
   try {
